+++ eslint.config.js
@@ -1,1 +1,5 @@
-{}
+{
+  "rules"; {
+    "@typescript-eslint/no-require-imports"; "error"
+  }
+}